Stick to bcrypt for passwords
In this commit we simplify the assumptions for the Password object,
constraining it to be of a type alias of string in order to have control
of the flow of data at compile time. We also add on top of it some
functionality to be sure to always serialize it as an hash.